products-management/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.5
djangorestframework==3.15.2
filelock==3.17.0
fonttools==4.55.8
fsspec==2024.12.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.70
ultralytics-thop==2.0.14
urllib3==2.3.0
